refactor: (#53) Main 공통 레이아웃 구조 정리
