+++ src/app/routes.tsx
@@ -1,4 +1,4 @@
-import MainPage from '@/pages/main/ui/MainPage';
+import MainPage from '@/pages/main/ui/layout/MainPage';
 import ProfilePage from '@/pages/profile/ui/ProfilePage';
 
 const routes = [
