+++ src/pages/main/ui/layout/MainHero.tsx
