+++ src/pages/main/ui/layout/MainTabSection.tsx
@@ -1,12 +1,11 @@
 import { Plus } from 'lucide-react';
-
-import type { MainTab } from '../model/types';
-import { mockRooms } from '../mocks/mockRooms';
-import MainBoardSection from './MainBoardSection';
-import MainLunchMenuSection from './MainLunchMenuSection';
-import MainRankingSection from './MainRankingSection';
-import RoomCard from './RoomCard';
-import RoomSummary from './RoomSummary';
+import type { MainTab } from '../../model/types';
+import { mockRooms } from '../../mocks/mockRooms';
+import MainBoardSection from '../MainBoardSection';
+import MainLunchMenuSection from '../MainLunchMenuSection';
+import MainRankingSection from '../MainRankingSection';
+import RoomCard from '../RoomCard';
+import RoomSummary from '../RoomSummary';
 
 interface MainTabSectionProps {
   activeTab: MainTab;
