+++ src/pages/profile/ui/ProfilePage.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import MainHeader from '@/pages/main/ui/MainHeader';
+import MainHeader from '@/pages/main/ui/layout/MainHeader';
 import { updateMyProfile } from '@/shared/api/users/profile';
 import { myProfileQueryOptions } from '@/shared/api/users/profileQueries';
 import { MBTI_OPTIONS, type MbtiType, type UserProfile } from '@/shared/types/profile';
